nc_close {ncdf4}	R Documentation

Close a netCDF File

Description

Closes an open netCDF file, which flushes any unwritten data to disk. Always close a netCDF file when you are done with it! You are risking data loss otherwise.

Usage

 nc_close( nc )

Arguments

nc
An object of class ncdf4 (as returned by either function nc_open or function nc_create.

Details

Data written to a netCDF file is cached in memory, for better performance. This data is only written out to disk when the file is closed. Therefore, always remember to close a netCDF file when done with it.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://dwpierce.com/software

See Also

nc_sync.

Examples

## Not run: nc <- nc_open("salinity.nc")
## Not run: data <- ncvar_get( nc )	# Read the "only" var in the file
## Not run: nc_close(nc)

[Package ncdf4 version 1.8 Index]