nc_redef {ncdf4}	R Documentation

Puts a netCDF file back into define mode

Description

Puts a netCDF that is not currently in define mode back into define mode.

Usage

 nc_redef( nc )

Arguments

nc
An object of class ncdf4 (as returned by either function nc_open(..., write=TRUE) or function nc_create, indicating what file to operate on.

Details

Typically, users will never need this function.

NetCDF files can be in "define mode", at which time dimensions and variables can be defined, or new attributes added to a file, or in "data mode", at which time data can be read from the file. This call puts a file that is currently in data mode back into define mode. This functionality is handled transparently by the ncdf4 library, so users will never need to call this unless they are doing advanced manipulations of netcdf files.

Note

The typical user will never need this call, nor will ever have to worry about "define mode" or "data mode". THIS CALL IS PROVIDED FOR ADVANCED USERS ONLY! If the user goes through this package's standard functional interface, the file will always automatically be set to whatever mode it needs to be in without the user having to do anything. For example, the call to write an attribute (ncatt_put) handles this automatically.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://dwpierce.com/software

See Also

nc_enddef.

Examples

# This function is for advanced useage only, and will never
# be needed by the typical users R code.

[Package ncdf4 version 1.8 Index]