nc_version {ncdf4}	R Documentation

Report version of ncdf4 library

Description

Returns a string that is the version number of the ncdf4 package.

Usage

 nc_version()

Details

Note that the returned value it is a string, not a floating point number.

Value

A string (not float) that is the version number of the ncdf4 package.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://dwpierce.com/software

[Package ncdf4 version 1.8 Index]